cleanup in variablenames and styling
