+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "i18n-ally.localesPaths": ["i18n", "messages", "src/i18n"]
+}
