+++ src/components/employeeCard/employeeCard.module.css
@@ -1,11 +1,10 @@
-/*---------CustomerCase consultantCard------*/
 .wrapper {
   display: flex;
   align-items: center;
   gap: 1rem;
 }
 
-.consultantImage {
+.employeeImage {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -17,7 +16,7 @@
   position: relative;
 }
 
-.consultantInfoWrapper {
+.employeeInfoWrapper {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
@@ -27,19 +26,17 @@
 .contactInfo {
   font-size: 12px;
   font-weight: 400;
-  line-height: 120%; /* 14.4px */
+  line-height: 120%;
   gap: 0.375rem;
 }
 
-.consultantName {
+.employeeName {
   font-size: 1rem;
   font-weight: 600;
-  line-height: 120%; /* 19.2px */
+  line-height: 120%;
 }
 
-/*---------ConsultantCardPeoplePage------*/
-
-.peopleWrapper {
+.employeeWrapper {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
@@ -49,7 +46,7 @@
   gap: var(--small, 6px);
 }
 
-.peopleConsultantInfoWrapper {
+.employeeInfoWrapper {
   display: flex;
   text-wrap: wrap;
   flex-direction: column;
@@ -59,7 +56,7 @@
   align-self: stretch;
 }
 
-.peopleConsultantImage {
+.employeeImage {
   display: block;
   height: 206px;
   min-width: 280px;
@@ -70,9 +67,7 @@
   border-radius: var(--medium, 12px);
 }
 
-/*---------Name_Competence-------------*/
-
-.peopleConsultantInfo {
+.employeeInfo {
   color: var(--stroke-tertiary, #5e5e5e);
   font-size: 1rem;
   font-style: normal;
@@ -80,29 +75,28 @@
   line-height: 120%;
 }
 
-.peopleConsultantName {
+.employeeName {
   color: var(--text-primary, #222424);
 }
 
-.peopleConsultantRole {
+.employeeRole {
   display: flex;
   flex-direction: row;
   overflow: visible;
   align-self: stretch;
 }
 
-.peopleConsultantRoleDot::after {
+.employeeRoleDot::after {
   content: "·";
   margin: 0 0.5rem;
 }
 
-.peopleConsultantRoleDot:last-child:after {
+.employeeRoleDot:last-child:after {
   content: "";
   margin: 0;
 }
-/*-------------Email_PhoneNR----------------*/
 
-.peopleContactInfo {
+.employeeContactInfo {
   font-weight: 400;
   line-height: 120%;
   gap: 0.375rem;
